Move input dialog code to a separate folder
