+++ packages-ui/roosterjs-react/lib/inputDialog/type/DialogItem.ts
