+++ packages-ui/roosterjs-react/lib/ribbon/component/buttons/insertLink.ts
@@ -1,5 +1,5 @@
 import RibbonButton from '../../type/RibbonButton';
-import showInputDialog from '../../../common/utils/showInputDialog';
+import showInputDialog from '../../../inputDialog/utils/showInputDialog';
 import { createLink } from 'roosterjs-editor-api';
 import { InsertLinkButtonStringKey } from '../../type/RibbonButtonStringKeys';
 import { QueryScope } from 'roosterjs-editor-types';
